+++ Telegram.BotAPI/Parameters/UpdatingMessages/DeleteBusinessMessagesParameters.cs
@@ -6,5 +6,5 @@ public sealed class DeleteBusinessMessagesParameters : ApiRequestParameters
 {
     public required string BusinessConnectionId {  get; init; }
 
-    public required IReadOnlyList<int> MessageIds { get; init; }
+    public required IReadOnlyList<long> MessageIds { get; init; }
 }
