+++ Telegram.BotAPI/Parameters/UpdatingMessages/ReadBusinessMessageParameters.cs
@@ -6,5 +6,5 @@ public sealed class ReadBusinessMessageParameters : ApiRequestParameters
 
     public required long ChatId { get; init; }
 
-    public required int MessageId { get; init; }
+    public required long MessageId { get; init; }
 }
