+++ Telegram.BotAPI/Types/AvailableTypes/MaybeInaccessibleMessage.cs
@@ -4,7 +4,7 @@ public abstract class MaybeInaccessibleMessage
 {
     public required virtual Chat Chat { get; init; }
 
-    public required virtual int MessageId { get; init; }
+    public required virtual long MessageId { get; init; }
 
-    public required virtual int Date { get; init; }
+    public required virtual long Date { get; init; }
 }
