+++ Telegram.BotAPI/Types/AvailableTypes/MessageReactionCountUpdated.cs
@@ -6,9 +6,9 @@ public sealed class MessageReactionCountUpdated
 {
     public required Chat Chat { get; init; }
 
-    public required int MessageId { get; init; }
+    public required long MessageId { get; init; }
 
-    public required int Date { get; init; }
+    public required long Date { get; init; }
 
     public required IReadOnlyList<ReactionCount> Reactions { get; init; }
 }
