+++ Telegram.BotAPI/Types/AvailableTypes/MessageReactionUpdated.cs
@@ -6,13 +6,13 @@ public sealed class MessageReactionUpdated
 {
     public required Chat Chat { get; init; }
 
-    public required int MessageId { get; init; }
+    public required long MessageId { get; init; }
 
     public User? User { get; init; }
 
     public Chat? ActorChat { get; init; }
 
-    public required int Date { get; init; }
+    public required long Date { get; init; }
 
     public required IReadOnlyList<ReactionType> OldReaction { get; init; }
